refactor: move ai stuff
