+++ src/utils/functions/ai/help-chat.ts
@@ -1,13 +1,13 @@
 import { ResponsesModel } from "openai/resources";
-import prisma from "../../init/database";
-import { getCommandData, getCommandKeys } from "../handlers/commandhandler";
-import { logger } from "../logger";
-import { isLockedOut } from "./captcha";
-import { getLevel, getPrestige } from "./economy/levelling";
-import { isEcoBanned, userExists } from "./economy/utils";
+import prisma from "../../../init/database";
+import { getCommandData, getCommandKeys } from "../../handlers/commandhandler";
+import { logger } from "../../logger";
+import { isLockedOut } from "../captcha";
+import { getLevel, getPrestige } from "../economy/levelling";
+import { isEcoBanned, userExists } from "../economy/utils";
+import { getLastCommand } from "../users/commands";
+import { getLastKnownUsername } from "../users/username";
 import openai, { buildPrompt, getDocsRaw } from "./openai";
-import { getLastCommand } from "./users/commands";
-import { getLastKnownUsername } from "./users/username";
 
 const MODEL: ResponsesModel = "gpt-5.4-nano";
 type ChatHistoryInput = { role: "user" | "assistant"; content: string };
