+++ src/utils/functions/supportrequest.ts
@@ -9,11 +9,11 @@ import { NypsiClient } from "../../models/Client";
 import { CustomEmbed } from "../../models/EmbedBuilders";
 import Constants from "../Constants";
 import { logger } from "../logger";
+import openai, { buildPrompt, getDocsRaw, prompt } from "./ai/openai";
 import { isLockedOut } from "./captcha";
 import { getLevel, getPrestige } from "./economy/levelling";
 import { isEcoBanned, userExists } from "./economy/utils";
 import { uploadImage } from "./image";
-import openai, { buildPrompt, getDocsRaw, prompt } from "./openai";
 import { getLastCommand } from "./users/commands";
 import { getLastKnownUsername } from "./users/username";
 import pAll = require("p-all");
